requirements.txt
fastapi
uvicorn
gunicorn
sqlalchemy
mysql-connector-python
python-jose
passlib[bcrypt]
